src/service/tools/requirements.txt
Django>=4.2
gunicorn
psycopg2-binary
stripe
python-dotenv
djangorestframework
django-cors-headers